# services/gateway/docs/swagger.yaml
basePath: /
definitions:
  ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  actionResponse:
    properties:
      state:
        type: string
    type: object
...
  friendRequestActionResponse:
    properties:
      request_id:
        type: string
      state:
        type: string
    type: object
  friendRequestResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      recipient_id:
        type: string
      responded_at:
        type: string
      sender_id:
        type: string
      status:
        type: string
    type: object
  friendRequestsResponse:
    properties:
      incoming:
        items:
          $ref: '#/definitions/friendRequestResponse'
        type: array
      outgoing:
        items:
          $ref: '#/definitions/friendRequestResponse'
        type: array
    type: object
...
  profileResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      handle:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  recommendationResponse:
    properties:
      common_friends:
        type: integer
      user_id:
        type: string
    type: object
  recommendationsResponse:
    properties:
      recommendations:
        items:
          $ref: '#/definitions/recommendationResponse'
        type: array
    type: object
...
  relationshipsResponse:
    properties:
      friends:
        items:
          type: string
        type: array
      subscribers:
        items:
          type: string
        type: array
      subscriptions:
        items:
          type: string
        type: array
    type: object
...
  setHandleRequest:
    properties:
      handle:
        type: string
    type: object
...
  updateProfileRequest:
    properties:
      bio:
        type: string
      display_name:
        type: string
    type: object
...
info:
  contact: {}
  description: Единый публичный HTTP-контракт Go-бэкенда.
  title: General Project API
  version: "1.0"
paths:
...
  /v1/profiles/me:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Публичные поля профиля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/updateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Обновить публичный профиль
      tags:
      - profiles
...
  /v1/profiles/me/handle:
    put:
      consumes:
      - application/json
      parameters:
      - description: Новый handle
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/setHandleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Назначить handle профиля
      tags:
      - profiles
...
  /v1/social/friend-requests:
    get:
      parameters:
      - default: all
        description: incoming, outgoing или all
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/friendRequestsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Получить заявки в друзья
      tags:
      - friend-requests
  /v1/social/friend-requests/{requestID}:
    delete:
      parameters:
      - description: UUID заявки
        in: path
        name: requestID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/friendRequestActionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Отменить заявку в друзья
      tags:
      - friend-requests
  /v1/social/friend-requests/{requestID}/accept:
    post:
      parameters:
      - description: UUID заявки
        in: path
        name: requestID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/friendRequestActionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Принять заявку в друзья
      tags:
      - friend-requests
  /v1/social/friend-requests/{requestID}/decline:
    post:
      parameters:
      - description: UUID заявки
        in: path
        name: requestID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/friendRequestActionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Отклонить заявку в друзья
      tags:
      - friend-requests
  /v1/social/friendships/{targetID}:
    delete:
      parameters:
      - description: UUID пользователя
        in: path
        name: targetID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/actionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Удалить из друзей
      tags:
      - friendships
    put:
      parameters:
      - description: UUID пользователя
        in: path
        name: targetID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/friendRequestActionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Отправить заявку в друзья
      tags:
      - friendships
  /v1/social/recommendations:
    get:
      parameters:
      - default: 10
        description: Количество результатов
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recommendationsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Получить рекомендации друзей
      tags:
      - social
  /v1/social/relationships:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/relationshipsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Получить социальные связи
      tags:
      - social
  /v1/social/subscriptions/{targetID}:
    delete:
      parameters:
      - description: UUID пользователя
        in: path
        name: targetID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/actionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Отписаться от пользователя
      tags:
      - subscriptions
    put:
      parameters:
      - description: UUID пользователя
        in: path
        name: targetID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/actionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Подписаться на пользователя
      tags:
      - subscriptions
swagger: "2.0"